feat: implementação da persistência de usuários e refatoração de herança
Implementação do método cadastrar em UsuarioDAO utilizando PreparedStatement e JDBC.

Uso de instanceof no DAO para identificar o tipo de conta e salvar corretamente no banco de dados.

Criação da lógica de classificação para padronizar os tipos de usuários do sistema.

Refatoração dos construtores de Cliente e Administrador para incluir o id, respeitando a herança de Usuario.

Manutenção das classes ClienteDAO e AdministradorDAO para futuras implementações específicas.

+++ src/model/Produto.java
@@ -22,7 +22,7 @@ public class Produto {
         return true;
     }
 
-    public boolean validacaoPreco(int preco){
+    public boolean validacaoPreco(double preco){
         if (preco <= 0){
             return false;
         }
